+++ core/src/cpp/include/synth/note.hpp
@@ -1,4 +1,4 @@
-#ifndef MC64K_STANDARD_TEST_HOST_SYNTH_NITE_HPP
+#ifndef MC64K_STANDARD_TEST_HOST_SYNTH_NOTE_HPP
     #define MC64K_STANDARD_TEST_HOST_SYNTH_NOTE_HPP
 
 /**
